app/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
crispy-bootstrap3==2022.1
dj-database-url==2.0.0
Django==4.2.1
django-crispy-forms==2.0
django-debug-toolbar==4.0.0
django-embed-video==1.4.10
django-filter==23.2
djangorestframework==3.14.0
easy-thumbnails==2.10
et-xmlfile==2.0.0
Flask==2.2.5
gunicorn==20.1.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.4
keyboard==0.13.5
MarkupSafe==3.0.2
MouseInfo==0.1.3
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
paperclip==2.6.1
Pillow==9.3.0
psycopg2-binary==2.9.6
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2023.3
pywhatkit @ https://github.com/bichocj/PyWhatKit/archive/refs/heads/master.zip
pywin32==308
requests==2.28.2
six==1.16.0
soupsieve==2.6
sqlparse==0.4.4
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.20
werkzeug==3.0.6
whitenoise==6.4.0
wikipedia==1.4.0
xlrd==2.0.1
